+++ .travis/setartname.sh
@@ -1,6 +1,1 @@
-if [[ $TRAVIS_TAG ]]
- then
-  export ARTIFACT_NAME="$TRAVIS_TAG";
-else
-  export ARTIFACT_NAME="$TRAVIS_BRANCH";
-fi
+export ARTIFACT_NAME="$(meteor npm run version --silent)"
